Rename message translators
