+++ source/modulo_new_core/modulo_new_core/translators/message_readers.py
